+++ src/main/kotlin/enums/Rank.kt
@@ -1,4 +1,9 @@
-enum class Rank(val countOfMatch: Int, val winningMoney: Int) {
+package enums
+
+enum class Rank(
+    val countOfMatch: Int,
+    val winningMoney: Int,
+) {
     FIRST(6, 2_000_000_000),
     SECOND(5, 30_000_000),
     THIRD(5, 1_500_000),
