+++ src/test/kotlin/LottoCalculatorTest.kt
@@ -1,7 +1,8 @@
-import domain.Lotto
-import domain.LottoCalculator
-import domain.WinningLotto
+import domain.model.Lotto
+import domain.model.WinningLotto
+import domain.service.LottoCalculator
 import domain.value.LottoNumber
+import enums.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
